Make the decl id a size_t
    We're using size_t in the printf format.

+++ parse/parse.h
@@ -150,7 +150,7 @@ struct Node {
             Op op;
             Type *type;
             int isconst;
-            long   did; /* for Ovar, we want a mapping to the decl id */
+            size_t did; /* for Ovar, we want a mapping to the decl id */
             size_t nargs;
             Node **args;
         } expr;
